negocios_push/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# ca-certificates==2020.1.1
certifi==2019.11.28
django==3.0.3
asgiref==3.2.3
pip==20.0.2
pytz==2019.3
setuptools==45.2.0
# sqlite==3.31.1
sqlparse==0.3.0
Image==1.5.28
pillow==7.0.0
vc==2018.7.10
# vs2015_runtime==14.16.27012
wheel==0.34.2
wincertstore==0.2
gunicorn==20.0.4
whitenoise==5.0.1
djangorestframework==3.11.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
six==1.14.0